KOP-657, Aanpassen in de code KOP-638, aud en iss  in de HTI token: Topic 007 en de DevGuide out of sync
 - Refactored all token validation
