+++ application/oauth_server/service.py
@@ -1,24 +1,24 @@
+#  Copyright (c) Stichting Koppeltaal 2021.
+#
+#  This Source Code Form is subject to the terms of the Mozilla Public
+#  License, v. 2.0. If a copy of the MPL was not distributed with this
+#  file, You can obtain one at https://mozilla.org/MPL/2.0/.
 import logging
 from base64 import urlsafe_b64decode
 from calendar import timegm
 from datetime import datetime, timezone
 from hashlib import sha256
-from typing import Dict, Any, Optional
 from uuid import uuid4
 
-import jwt as pyjwt
 from authlib.jose import jwt
 from flask import request, current_app
-from jwt import InvalidSignatureError, DecodeError
 
 from application.jwks.service import keypair_service
-from application.oauth_server.jwks_client import CacheHeaderPyJWKClient
-from application.oauth_server.model import Oauth2Token, SmartService, SmartServiceStatus
+from application.oauth_server.model import Oauth2Token
 
 logger = logging.getLogger('oauth_service')
 logger.setLevel(logging.DEBUG)
 
-consumed_jti_tokens = []
 
 LAUNCH_SCOPE_DEFAULT = 'launch openid fhirUser'
 LAUNCH_SCOPE_ALLOWED = [['launch'], ['launch', 'openid', 'fhirUser']]
@@ -103,264 +103,10 @@ class TokenService:
         return jwt.encode(header, payload, private_key).decode('ascii')
 
 
-class Oauth2ClientCredentialsService:
-    """
-    This class validates tokens issued by a registered application as smart_service.
-    """
-    consumed_jti_tokens = []
-
-    def verify_and_get_token(self, encoded_token, auth_client_id=None):
-
-        logger.debug(f'Received encoded token: {encoded_token}')
-        try:
-            unverified_decoded_jwt = pyjwt.decode(encoded_token, options={"verify_signature": False})
-        except DecodeError as e:
-            logger.warning(f"Failed to decode JWT: {e}")
-            return
-
-        if 'iss' not in unverified_decoded_jwt:
-            logger.warning("JWT doesn't contain a iss value")
-            return
-
-        client_id = unverified_decoded_jwt['iss']
-
-        logger.debug(
-            f'Verifying received token with body: {unverified_decoded_jwt} and header: {pyjwt.get_unverified_header(encoded_token)}')
-
-        smart_service: SmartService = self.get_smart_service(unverified_decoded_jwt)
-
-        if not smart_service or smart_service.status != SmartServiceStatus.APPROVED:
-            logger.warning(
-                f"Discontinuing request for client_id {client_id}, smart service not found or status not approved")
-            return
-
-        if 'jti' not in unverified_decoded_jwt:
-            logger.warning("JWT doesn't contain a jti value")
-            return
-
-        # FIXME: This should be checked against a shared cache like Redis to support multiple instances / ease of rebooting the  application
-        if unverified_decoded_jwt['jti'] in consumed_jti_tokens:
-            logger.warning(f"JWT is being replayed - jti {unverified_decoded_jwt['jti']} is already consumed")
-            return
-
-        if 'aud' in unverified_decoded_jwt:
-            aud = unverified_decoded_jwt['aud']
-            if not self._verify_aud(aud, auth_client_id):
-                logger.warning(f'Failed to verify the aud value ({aud}) of the JWT token.')
-                return
-        else:
-            logger.warning("JWT doesn't contain a aud value")
-            return
 
 
-        consumed_jti_tokens.append(unverified_decoded_jwt['jti'])
 
-        try:
-            if smart_service.jwks_endpoint:
-                return self.decode_with_jwks(smart_service, encoded_token)
-            elif smart_service.public_key:
-                return self.decode_with_public_key(smart_service, encoded_token)
-            else:
-                logger.error(f"No JWKS or Public Key found on smart service with client_id {client_id}", )
-        except InvalidSignatureError as ise:
-            logger.warning(f"Invalid signature for client_id {client_id}, exception {ise}")
-            return
-        except Exception as e:
-            logger.warning(
-                f"Something went wrong whilst trying to decode the JWT for client_id {client_id}, exception {e}")
-            return
-
-    def decode_with_jwks(self, smart_service, encoded_token) -> Dict[str, Any]:
-        logger.info(f'Fetching endpoint: "{smart_service.jwks_endpoint}" for client_id: {smart_service.client_id}')
-        jwks_client = CacheHeaderPyJWKClient(smart_service.jwks_endpoint)
-        signing_key = jwks_client.get_signing_key_from_jwt(encoded_token)
-        decoded_jwt = pyjwt.decode(encoded_token, signing_key.key,
-                                   algorithms=current_app.config['OIDC_SMART_CONFIG_SIGNING_ALGS'],
-                                   options={'verify_aud': False})
-        logger.info(f'JWT for client_id {smart_service.client_id} is decoded by JWKS - valid key')
-        return decoded_jwt
-
-    def decode_with_public_key(self, smart_service, encoded_token) -> Dict[str, Any]:
-
-        public_key = smart_service.public_key
-
-        if not public_key.startswith("-----BEGIN PUBLIC KEY-----"):
-            logger.debug(
-                f"public key for client_id {smart_service.client_id} didn't contain -----BEGIN PUBLIC KEY-----, injecting start and end tags")
-            public_key = f'-----BEGIN PUBLIC KEY-----\n{public_key}\n-----END PUBLIC KEY-----'
-
-        decoded_jwt = pyjwt.decode(encoded_token, public_key,
-                                   algorithms=["RS512"],  ## TODO: check with spec
-                                   options={'verify_aud': False})
-
-        logger.info(f'JWT for client_id {smart_service.client_id} is decoded by PUBLIC KEY - valid key')
-        return decoded_jwt
-
-    def get_smart_service(self, unverified_decoded_jwt):
-        issuer = unverified_decoded_jwt['iss']
-
-        smart_service = SmartService.query.filter_by(client_id=issuer).first()
-        logger.info(f'Matched issuer {issuer} to smart service {smart_service}')
-
-        return smart_service
-
-    def _verify_aud(self, aud: str, auth_client_id):
-        """
-        The client credential token can either be a HTI token or a client auth token
-        :param aud: the value from the auth field
-        :param auth_client_id: the client id set as the issuer of the token
-        :return: True if the aud is valid
-        """
-        token_endpoint = self._get_token_endpoint()
-        if aud.startswith("Device/"): # Check if the audience is the Device/123 reference.
-            smart_service: SmartService = SmartService.query.filter_by(fhir_store_device_id=aud).first()
-            if smart_service and smart_service.client_id == auth_client_id:
-                return True
-
-            logger.warning(
-                f'Unable to find a smart service {aud} with client id {auth_client_id}')
-            return False
-        elif token_endpoint == aud:  # Check if the audience is the auth service itself.
-            return True
-
-        logger.warning(f'Unexpected aud in token: {aud}')
-        return False
-
-    def _get_token_endpoint(self):
-        return request.root_url + 'token'
-
-
-class ServerOauth2ClientCredentialsService():
-    def verify_and_get_token(self, encoded_token):
-        logger.debug(f'Received encoded token: {encoded_token}')
-        unverified_decoded_jwt = pyjwt.decode(encoded_token, options={"verify_signature": False})
-        if 'iss' not in unverified_decoded_jwt:
-            logger.warning("JWT doesn't contain a iss value")
-            return
-        iss = unverified_decoded_jwt['iss']
-        if iss != request.url_root:
-            logger.warning("JWT doesn't contain self as iss value")
-            return
-
-        if 'jti' not in unverified_decoded_jwt:
-            logger.warning("JWT doesn't contain a jti value")
-            return
-
-        # FIXME: This should be checked against a shared cache like Redis to support multiple instances / ease of rebooting the  application
-        if unverified_decoded_jwt['jti'] in consumed_jti_tokens:
-            logger.warning(f"JWT is being replayed - jti {unverified_decoded_jwt['jti']} is already consumed")
-            return
-
-        consumed_jti_tokens.append(unverified_decoded_jwt['jti'])
-
-        try:
-            return self.decode_with_own_key(encoded_token)
-
-        except InvalidSignatureError as ise:
-            logger.warning(f"Invalid signature validating own key, exception {ise}")
-            return
-        except Exception as e:
-            logger.warning(
-                f"Something went wrong whilst trying to decode the JWT, exception {e}")
-            return
-
-    def decode_with_own_key(self, encoded_token):
-        _, public_key = keypair_service.get_keypair()
-        decoded_jwt = pyjwt.decode(encoded_token, public_key.as_pem(),
-                                   algorithms=current_app.config[
-                                       'OIDC_SMART_CONFIG_SIGNING_ALGS'],
-                                   options={'verify_aud': False})  # TODO: check if correct
-
-        logger.info(f'JWT signed by self is decoded by JWKS - valid key')
-        return decoded_jwt
-
-
-class SmartHtiOnFhirService:
-    def __init__(self):
-        pass
-
-    @staticmethod
-    def validate_and_parse_launch_scope(scope: str, encoded_token: str):
-        scopes = scope.split()
-        for allowed_set in LAUNCH_SCOPE_ALLOWED:
-            if set(allowed_set) == set(scopes):
-                return scopes
-
-        return None
-
-    def validate_launch_token(self, encoded_token: str) -> Optional[Dict[str, Any]]:
-        unverified_decoded_jwt = pyjwt.decode(encoded_token, options={"verify_signature": False})
-        issuer = unverified_decoded_jwt['iss']
-
-        if not unverified_decoded_jwt['jti']:
-            logger.warning("JWT doesn't contain a jti value")
-            return
-
-        # FIXME: This should be checked against a shared cache like Redis to support multiple instances / ease of rebooting the  application
-        if unverified_decoded_jwt['jti'] in consumed_jti_tokens:
-            logger.warning(f"JWT is being replayed - jti {unverified_decoded_jwt['jti']} is already consumed")
-            return
-
-        consumed_jti_tokens.append(unverified_decoded_jwt['jti'])
-
-        smart_service: SmartService = self.get_smart_service(issuer)
-        if not smart_service:
-            logger.warning(f"Cannot find smart_service for issuer: {issuer}")
-            return
-
-        try:
-            if smart_service.jwks_endpoint:
-                return self.decode_with_jwks(smart_service, encoded_token)
-            elif smart_service.public_key:
-                return self.decode_with_public_key(smart_service, encoded_token)
-            else:
-                logger.error(f"No JWKS or Public Key found on smart service with client_id {issuer}")
-        except InvalidSignatureError as ise:
-            logger.warning(f"Invalid signature for client_id {issuer}, exception {ise}")
-            return
-        except Exception as e:
-            logger.warning(
-                f"Something went wrong whilst trying to decode the JWT for client_id {issuer}, exception {e}")
-            return
-
-    def decode_with_jwks(self, smart_service, encoded_token):
-        logger.info(f'Fetching endpoint: "{smart_service.jwks_endpoint}" for client_id: {smart_service.client_id}')
-        jwks_client = CacheHeaderPyJWKClient(smart_service.jwks_endpoint)
-        signing_key = jwks_client.get_signing_key_from_jwt(encoded_token)
-        decoded_jwt = pyjwt.decode(encoded_token, signing_key.key,
-                                   algorithms=current_app.config[
-                                       'OIDC_SMART_CONFIG_SIGNING_ALGS'],
-                                   options={'verify_aud': False})
-
-        logger.info(f'JWT for client_id {smart_service.client_id} is decoded by JWKS - valid key')
-        return decoded_jwt
-
-    def decode_with_public_key(self, smart_service, encoded_token):
-
-        public_key = smart_service.public_key
-
-        if not public_key.startswith("-----BEGIN PUBLIC KEY-----"):
-            logger.debug(
-                f"public key for client_id {smart_service.client_id} didn't contain -----BEGIN PUBLIC KEY-----, injecting start and end tags")
-            public_key = '-----BEGIN PUBLIC KEY-----\n' + public_key + '\n-----END PUBLIC KEY-----'
-        decoded_jwt = pyjwt.decode(encoded_token, public_key,
-                                   algorithms=["RS512"],
-                                   options={'verify_aud': False})
-
-        logger.info(f'JWT for client_id {smart_service.client_id} is decoded by PUBLIC KEY - valid key')
-        return decoded_jwt
-
-    def get_smart_service(self, client_id):
-        smart_service = SmartService.query.filter_by(client_id=client_id).first()
-        logger.info(f'Matched issuer {client_id} to smart service {smart_service}')
-
-        return smart_service
-
-
-smart_hti_on_fhir_service = SmartHtiOnFhirService()
 token_service = TokenService()
-oauth2_client_credentials_service = Oauth2ClientCredentialsService()
-server_oauth2_service = ServerOauth2ClientCredentialsService()
 token_authorization_code_service = TokenAuthorizationCodeService()
 
 """
