+++ test/test_oauth_flows.py
@@ -83,7 +83,7 @@ def smart_service_client(testing_app: FlaskClient, client_key: Key, client_id: s
                                         client_id=client_id,
                                         status=SmartServiceStatus.APPROVED,
                                         public_key=public_key_bytes.decode('utf8'),
-                                        fhir_store_device_id=str(uuid4()),
+                                        fhir_store_device_id=f'Device/{client_id}',
                                         patient_idp=identity_provider.id,
                                         practitioner_idp=identity_provider.id)
     db.session.add(smart_service_client)
@@ -98,6 +98,7 @@ def smart_service_portal(portal_key: Key, portal_id: str, identity_provider: Ide
                                         client_id=portal_id,
                                         status=SmartServiceStatus.APPROVED,
                                         public_key=public_key_bytes.decode('utf8'),
+                                        fhir_store_device_id=f'Device/{portal_id}',
                                         patient_idp=identity_provider.id,
                                         practitioner_idp=identity_provider.id)
     db.session.add(smart_service_portal)
@@ -112,7 +113,7 @@ def smart_service_custom_idp(identity_provider, client_key: Key, client_id_idp: 
                                         client_id=client_id_idp,
                                         status=SmartServiceStatus.APPROVED,
                                         public_key=public_key_bytes.decode('utf8'),
-                                        fhir_store_device_id=str(uuid4()),
+                                        fhir_store_device_id=f'Device/{client_id_idp}',
                                         patient_idp=identity_provider.id,
                                         practitioner_idp=identity_provider.id)
     print("custom idp met client_id: ", client_id_idp)
@@ -298,7 +299,7 @@ def test_authorization_code_happy_without_verifier(mock1, mock2, testing_app: Fl
             'redirect_uri': allowed_redirect.url,
             'aud': testing_app.application.config.get('FHIR_CLIENT_SERVERURL'),
             'client_id': client_id,
-            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id),
+            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id, smart_service_client.fhir_store_device_id),
             'state': state}
     authorize_resp = testing_app.get(f'/oauth2/authorize?{urlencode(data)}')
     idp_code = str(uuid4())
@@ -344,7 +345,7 @@ def test_authorization_code_with_custom_idp(mock_get, mock_post, testing_app: Fl
             'redirect_uri': allowed_redirect.url,
             'aud': testing_app.application.config.get('FHIR_CLIENT_SERVERURL'),
             'client_id': smart_service_custom_idp.client_id,
-            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id),
+            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id, smart_service_custom_idp.fhir_store_device_id),
             'state': module_state}
 
     ## AUTHORIZE
@@ -397,7 +398,7 @@ def test_authorization_with_invalid_redirect_uri(mock_get, testing_app: FlaskCli
             'redirect_uri': "https://invalid.redirect.url",
             'aud': testing_app.application.config.get('FHIR_CLIENT_SERVERURL'),
             'client_id': smart_service_client.client_id,
-            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id),
+            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id, smart_service_client.fhir_store_device_id),
             'state': module_state}
 
     authorize_resp = testing_app.get(f'/oauth2/authorize?{urlencode(data)}')
@@ -415,14 +416,15 @@ def test_authorization_with_zero_redirect_uri(mock_get, mock_post, testing_app: 
                                               patient_id: str,
                                               resource_id: str,
                                               smart_service_portal: SmartService,
+                                              smart_service_client: SmartService,
                                               custom_idp_location: str):
 
     module_state = str(uuid4())
     data = {'scope': 'launch fhirUser openid',
             'redirect_uri': "https://whatever.redirect.url",
             'aud': testing_app.application.config.get('FHIR_CLIENT_SERVERURL'),
-            'client_id': smart_service_portal.client_id,
-            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id),
+            'client_id': smart_service_client.client_id,
+            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id, smart_service_client.fhir_store_device_id),
             'state': module_state}
 
     authorize_resp = testing_app.get(f'/oauth2/authorize?{urlencode(data)}')
@@ -440,6 +442,7 @@ def test_multiple_configured_redirect_uris(mock_get, mock_post, testing_app: Fla
                                               patient_id: str,
                                               resource_id: str,
                                               smart_service_portal: SmartService,
+                                              smart_service_client: SmartService,
                                               custom_idp_location: str,
                                               allowed_redirect: AllowedRedirect):
 
@@ -447,8 +450,8 @@ def test_multiple_configured_redirect_uris(mock_get, mock_post, testing_app: Fla
     data = {'scope': 'launch fhirUser openid',
             'redirect_uri': allowed_redirect.url,
             'aud': testing_app.application.config.get('FHIR_CLIENT_SERVERURL'),
-            'client_id': smart_service_portal.client_id,
-            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id),
+            'client_id': smart_service_client.client_id,
+            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id, smart_service_client.fhir_store_device_id),
             'state': module_state}
 
     authorize_resp = testing_app.get(f'/oauth2/authorize?{urlencode(data)}')
@@ -478,7 +481,7 @@ def test_authorization_code_happy_with_verifier(mock1, mock2, testing_app: Flask
             'code_challenge_method': 'S256',
             'aud': testing_app.application.config.get('FHIR_CLIENT_SERVERURL'),
             'client_id': client_id,
-            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id),
+            'launch': _hti_token(testing_app, portal_key, portal_id, user_id, patient_id, resource_id, smart_service_client.fhir_store_device_id),
             'state': state}
     authorize_resp = testing_app.get(f'/oauth2/authorize?{urlencode(data)}')
     idp_code = str(uuid4())
