+++ test/utils.py
@@ -23,12 +23,12 @@ def _client_assertion(testing_app: FlaskClient, client_key: Key, client_id: str,
     return jwt.encode(payload, get_private_key_as_pem(client_key), algorithm="RS512")
 
 
-def _hti_token(testing_app: FlaskClient, portal_key: Key, portal_id: str, user_id: str, patient_id: str, resource_id: str):
+def _hti_token(testing_app: FlaskClient, portal_key: Key, portal_id: str, user_id: str, patient_id: str, resource_id: str, audience:str):
     payload = {'iss': portal_id,
                'jti': str(uuid4()),
                'sub': user_id,
                'resource': resource_id,
-               'aud': testing_app.application.config['OIDC_SMART_CONFIG_TOKEN_ENDPOINT']}
+               'aud': audience}
 
     if patient_id:
         payload['patient'] = patient_id
